test:修改react-native-multiple-select测试demo
